requirements.txt
numpy
matplotlib
pillow
rasterio
scikit-learn
tqdm
seaborn
torch
torchvision
torchaudio